Extracts some methods to constants and adds axis labels and chart title for results
